tweak costs for bike paths
This lowers the cost for routing on a Way that's part of a cycling
Relation, but increases the cost for routing on a bike path/shared
lane.

It has a small effect on edges that were already both bikeways and
part of a cycling relation - for instance, it takes the Vera Katz
Esplanade from 0.18 to 0.26 which doesn't change routing much, if at
all.

But it has a large (and desireable) effect on edges that are not
counted as bikeways, but are part of a cycling Relation - for
instance, Grebbedijk, outside of Arnhem, NL, goes from a 0.48 to a
0.39.

These numbers work pretty well in my testing so far, but are likely to
change again soon.